feat: enhance mission tracking with progress display and status updates
